requirements.txt
amqp==5.2.0
anyio==3.7.1
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
async-timeout==4.0.3
billiard==4.2.0
celery==5.3.6
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
cryptography==41.0.4
cssselect2==0.7.0
dj-database-url==2.2.0
Django==5.0.2
django-admin-extra-buttons==1.5.7
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-ckeditor==6.7.0
django-cors-headers==4.2.0
django-crontab==0.7.1
django-debug-toolbar==4.2.0
django-filter==23.5
django-jazzmin==2.6.0
django-js-asset==2.1.0
django-rest-framework-nested==0.0.1
django-timezone-field==6.1.0
djangorestframework==3.14.0
docopt==0.6.2
drf-yasg==1.21.7
exceptiongroup==1.1.3
gunicorn==21.2.0
h11==0.14.0
html5lib==1.1
httpcore==0.17.3
httpx==0.24.1
idna==3.4
inflection==0.5.1
kombu==5.3.6
lxml==4.9.3
num2words==0.5.13
onesignal-sdk==2.0.0
oscrypto==1.3.0
packaging==23.1
pdfkit==1.0.0
Pillow==10.0.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.7
pycodestyle==2.11.0
pycparser==2.21
pyHanko==0.20.1
pyhanko-certvalidator==0.24.1
pypdf==3.16.4
pypng==0.20220715.0
python-bidi==0.4.2
python-crontab==3.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
qrcode==7.4.2
redis==5.0.3
reportlab==3.6.13
requests==2.31.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
svglib==1.5.1
tinycss2==1.2.1
tomli==2.0.1
typing_extensions==4.7.1
tzdata==2024.1
tzlocal==5.1
uritemplate==4.1.1
uritools==4.0.2
urllib3==2.0.4
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
xhtml2pdf==0.2.11
